% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanSites.R
\name{CleanSites}
\alias{CleanSites}
\title{Clean and Process Site Data for Spatial Analysis}
\usage{
CleanSites(x)
}
\arguments{
\item{x}{A simple features (SF) object representing site locations.}
}
\value{
This function does not return a value. Instead, it assigns three spatial layers to the global environment:
\itemize{
  \item `NewDivides`: A simple features (SF) object representing divides, transformed to CRS EPSG:4269.
  \item `dangermond_sites`: The intersection of the provided site locations with the divides.
  \item `dangermond_sites`: The result of joining `dangermond_sites` with `NewDivides`.
}
}
\description{
This function creates and processes spatial layers for divides, flow paths, and sites that intersect with divides. It reads spatial data from a GeoPackage and performs spatial operations such as intersection and joining.
}
\details{
The `CleanSites` function performs the following steps:
\itemize{
  \item Reads spatial data from the `nextgen_hydorfabric.gpkg` GeoPackage, specifically the `divides` layer.
  \item Transforms the CRS of the `divides` layer to EPSG:4269 and stores it in a variable `NewDivides`.
  \item Computes the intersection of the input site data (`x`) with `NewDivides` and stores the result in `dangermond_sites`.
  \item Joins the intersected sites (`dangermond_sites`) with `NewDivides` and updates the `dangermond_sites` variable.
}
}
