% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanWells.R
\name{CleanWells}
\alias{CleanWells}
\title{Clean and Process Well Water Level Data}
\usage{
CleanWells(x, y)
}
\arguments{
\item{x}{A data frame containing well water level data. Defaults to `all_combined_data`.}

\item{y}{A spatial data frame containing basin information. Defaults to `ActiveDivides`.}
}
\value{
A data frame with monthly averages of the change in well water levels for each basin. The output includes:
\describe{
  \item{Divide}{The divide ID.}
  \item{date}{The first day of the month for which the average is calculated.}
  \item{mean_change_mm}{The mean change in well water level in millimeters.}
}
}
\description{
This function filters, transforms, and processes well water level data from a combined dataset. It prepares the data for mapping and analysis by calculating changes in well water levels, joining with basin information, and summarizing the data on a monthly basis.
}
