% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_clean.R
\name{model_clean}
\alias{model_clean}
\title{Clean and Process Model Data}
\usage{
model_clean(x)
}
\arguments{
\item{x}{A data frame acquired from the `CABCMParquetRead()` or `TerraClimParquetRead()` function. This data frame should include columns for `divide_id`, `var`, `date`, `value`, and `source`.}
}
\value{
A data frame containing the combined and cleaned data. The dataset includes:
\itemize{
  \item Data points where `var` is not "str", including columns `divide_id`, `var`, `date`, `value`, and `source`.
  \item Calculated changes in soil moisture storage where `var` is "str", which is computed as the day-to-day difference in `value`.
  \item The combined data with appropriate renaming and omission of missing values.
}
}
\description{
This function processes the model data by separating soil moisture storage from the dataset, calculating changes in storage, and combining this information back with the original data.
}
\details{
The `model_clean` function performs the following steps:
\itemize{
  \item Filters out data points where `var` is not "str" and retains only those columns needed.
  \item For data points where `var` is "str", calculates the day-to-day change in soil moisture storage.
  \item Renames the calculated changes and combines this cleaned data with the original dataset.
}
}
