% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSeason.R
\name{PlotSeason}
\alias{PlotSeason}
\title{Plot Seasons}
\usage{
PlotSeason(SeasonData, season_name)
}
\arguments{
\item{SeasonData}{A spatial dataframe containing seasonal data with columns such as `percent_error` and `type`. This data should be output from the `process_model_data()` function.}

\item{season_name}{A character string representing the name of the season (e.g., "Winter", "Spring", "Summer", "Fall"). This will be used as the title of the plot.}
}
\value{
A `ggplot` object displaying the percent error across the spatial features of the Dangermond Preserve for the specified season.
}
\description{
This function generates standardized plots for the Dangermond Preserve based on seasonal data. The plots visualize the percent error across different spatial features for a given season. This function is designed to be used within the `GridSeasons()` function to arrange the plots in a grid layout.
}
\details{
The `PlotSeason` function performs the following operations:
\itemize{
  \item **Plot Creation**: Generates a ggplot object with a map tile background, using OpenStreetMap data.
  \item **Coordinate System**: Uses the spatial coordinate reference system `st_crs(26910)`.
  \item **Visualization**: Plots the spatial data with `percent_error` values visualized using a gradient color scale from white to dark red.
  \item **Title**: Sets the plot title to the provided `season_name`.
}
}
