% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_divide_ids.R
\name{process_divide_ids}
\alias{process_divide_ids}
\title{Process Data for a List of Divide IDs}
\usage{
process_divide_ids(divide_ids, bucket_name, prefix, region)
}
\arguments{
\item{divide_ids}{A character vector containing the divide IDs to process.}

\item{bucket_name}{A character string representing the name of the S3 bucket.}

\item{prefix}{A character string representing the prefix path in the S3 bucket.}

\item{region}{A character string representing the AWS region. Default is `"us-west-2"`.}
}
\value{
A data frame containing the combined data for all the divide IDs.
}
\description{
This function processes data for a list of divide IDs by selecting the corresponding basins and reading the associated parquet files from S3. The data is combined into a single data frame.
}
