% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet_files.R
\name{read_parquet_files}
\alias{read_parquet_files}
\title{Read Parquet Files from S3}
\usage{
read_parquet_files(site_names, bucket_name, prefix, region = "us-west-2")
}
\arguments{
\item{site_names}{A character vector containing site names for which to read parquet files.}

\item{bucket_name}{A character string representing the name of the S3 bucket.}

\item{prefix}{A character string representing the prefix path in the S3 bucket.}

\item{region}{A character string representing the AWS region. Default is `"us-west-2"`.}
}
\value{
A data frame containing the combined data from the parquet files, grouped by date.
}
\description{
This function lists and reads parquet files from an S3 bucket for a given set of site names. The data is processed to group by date and summarize numeric columns.
}
