% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_basin.R
\name{select_basin}
\alias{select_basin}
\title{Select Basin}
\usage{
select_basin(divide = "")
}
\arguments{
\item{divide}{A character string representing the `divide_id` to filter by. Default is an empty string.}
}
\value{
A data frame containing the sites (`All_Sites_sf`) that are within the specified basin.
}
\description{
This function filters a dataset of divides (`NewDivides`) by a specified `divide_id` and selects the corresponding basin. It then finds the sites within that basin from the dataset `All_Sites_sf`.
}
