% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteData.R
\name{SiteData}
\alias{SiteData}
\title{Import and Clean Site Data for Spatial Analysis}
\usage{
SiteData(x = NULL)
}
\arguments{
\item{x}{A data frame containing site location data with columns for station IDs, elevation, site names, and coordinates. If `NULL`, the function will load the `SiteLocations` dataset from the package.}
}
\value{
A simple features (SF) object with site data including coordinates and spatial reference.
}
\description{
This function reads site location data, either provided as a data frame or loaded from the included package dataset, cleans it, and converts it into a simple features (SF) object for spatial analysis.
}
\details{
The `SiteData` function performs the following steps:
\itemize{
  \item If no input data is provided (`x = NULL`), it loads the `SiteLocations` dataset from the `TNCDangermond` package.
  \item Converts `station_id_dendra`, `elev`, and `SiteName` columns to character type.
  \item Cleans up the `station_id_dendra` and `elev` columns, handling missing values and inconsistencies.
  \item Transforms `SiteName` to replace spaces with underscores.
  \item Filters rows where `SiteName` starts with "Dangermond" and excludes those containing "Weather".
  \item Converts the cleaned data to an SF object with coordinates defined by `lon` and `lat`, and sets the coordinate reference system (CRS) to EPSG:4269.
}
}
