% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerraClimParquetRead.R
\name{TerraClimParquetRead}
\alias{TerraClimParquetRead}
\title{Read and Combine Terra Climate Parquet Files into a Single Data Frame}
\usage{
TerraClimParquetRead(sub = "terraclim")
}
\arguments{
\item{sub}{A character string specifying the subfolder within the base folder to look for Parquet files. Default is "terraclim".}
}
\value{
A data frame containing the combined data from all Parquet files in the specified subfolder. The `var` column is standardized with values "run" for "q" and "str" for "soil".
}
\description{
This function reads all Parquet files from a specified subfolder related to Terra Climate data and combines them into a single data frame. It also performs some specific transformations on the `var` column to standardize variable names.
}
\details{
The `TerraClimParquetRead` function performs the following steps:
\itemize{
  \item Lists all Parquet files in the specified subfolder within the base folder `"/data/water_balance/"`.
  \item Reads each Parquet file into a data frame and stores them in a list.
  \item Combines all data frames in the list into a single data frame with an additional column `file` indicating the file from which each row was read.
  \item Transforms the `var` column to standardize variable names, changing "q" to "run" and "soil" to "str".
}
}
