% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen_model_data.R
\name{widen_model_data}
\alias{widen_model_data}
\title{Widen Model Data and Calculate Error}
\usage{
widen_model_data(data)
}
\arguments{
\item{data}{A data frame containing model output. The data should have columns for `var`, `value`, `ppt`, `aet`, `run`, `str`, and optionally `rch`.}
}
\value{
A data frame in wide format with an additional column `ERR`, representing the calculated error.
}
\description{
This function processes the output of a cleaned model dataset by pivoting the data to a wide format and calculating the error (ERR) based on available variables. The error is calculated as the difference between `ppt`, `aet`, and other relevant variables such as `rch`, `run`, and `str`.
}
\details{
The function performs the following steps:
\itemize{
  \item **Pivot Wider**: Converts the data to wide format with separate columns for each variable.
  \item **Error Calculation**: Calculates the error (`ERR`) as `ppt - aet - rch - run - str` when `rch` is available, or `ppt - aet - run - str` when it's not.
  \item **Filter**: Filters out rows where `divide_id` is equal to `"cat-351"`.
}
}
\note{
Ensure that the input data contains the required columns: `var`, `value`, `ppt`, `aet`, `run`, `str`, and optionally `rch`.
}
