% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCorrMap.R
\name{ModelCorrMap}
\alias{ModelCorrMap}
\title{Create a Map of Correlations Between Well Water Level and Model Recharge}
\usage{
ModelCorrMap(x, y = ModelScatterPlot, z = NewDivides)
}
\arguments{
\item{x}{A data frame resulting from the `Correlation()` function, containing:
\describe{
  \item{Divide}{The divide ID.}
  \item{correlation}{The correlation coefficient between `mean_change_mm` and `deltaSTR`.}
}}

\item{y}{A `ggplot` object produced by the `ModelScatter` function, representing the scatter plot
of well water level change vs. model recharge.}

\item{z}{A spatial data frame (sf object) with divide information, including the `divide_id` column}
}
\value{
A combined `ggplot` object with the correlation map and scatter plot side by side.
}
\description{
This function generates a map visualizing the correlation between well water level changes and
model-reported recharge. It colors divides based on whether the correlation is positive or negative.
It also combines this map with a scatter plot showing the relationship between well water levels and
model recharge.
}
\details{
The function takes in the output from the `Correlation()` function and uses it to create a map that
highlights divides with positive and negative correlations. It adds labels and includes various map
features like scale and north arrow. The combined map and scatter plot provide a comprehensive view
of the correlation analysis.
}
